requirements.txt
flask
opencv-python
google-genai
dotenv
numpy
